Refactoring minor things
